Modify tool to only support normalizing a heap dump (clearing field values)
This is useful to simplify comparison of heap dumps.
Heap dumps may contain caches that we cannot clear before creating a heap dump.
By zeroing cache fields in the heap dump, we can act as if the cache was empty
for comparison purposes.

We do not support String sanitization anymore.

We update README.md to reflect the new behavior/usage of the tool.